Rewrote from scratch to be smaller and simpler.
This rewrite is intended to breathe some new life into Relax by giving
it a narrower focus and making use of libraries that weren't around when
Relax was originally written. It is effectively the reusable bits from
the BreweryDB and PewPew Gems extracted into modules that can be used to
more quickly and easily build web service wrapper libraries.